Just various little things and checks that make life easier, especially if something goes wrong. Minions must be direct subdirectories of parent, and there can be no parent hierarchy. This is implicit ATM. Correct and simplify construction of CFLAGS, LDFLAGS and friends.

+++ src/mulle-bootstrap-zombify.sh
@@ -153,13 +153,15 @@ _bury_zombie()
 
    if [ -d "${stashdir}" ]
    then
-      is_minion_bootstrap_project "${stashdir}" && internal_fail "Would bury minion \"${stashdir}\" which is wrong"
+      if ! is_minion_bootstrap_project "${stashdir}"
+      then
+         _bury_stash "${reposdir}" "${name}" "${stashdir}"
 
-      _bury_stash "${reposdir}" "${name}" "${stashdir}"
-
-      exekutor rm ${COPYMOVEFLAGS} "${zombie}" >&2
-      exekutor rm ${COPYMOVEFLAGS} "${reposdir}/${name}" >&2
-
+         exekutor rm ${COPYMOVEFLAGS} "${zombie}" >&2
+         exekutor rm ${COPYMOVEFLAGS} "${reposdir}/${name}" >&2
+      else
+         log_fluff "Do not bury minion \"${stashdir}\""
+      fi
    else
       log_fluff "Zombie \"${stashdir}\" vanished or never existed ($PWD)"
    fi
